codec: Adapt codec::decoder methods to new state transitions
